+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add: documentation for the Run Custom Code object
 - Add: docuemntation for the Data Product Registry object
 - Add: docuemntation for the Data Product Marketplace object
+- Add: get-formatted-datetime utility function
 - Change: Provided updates to the global.css file to support special cases from the Data Product Marketplace object
 
 ## SAS Portal Framework for SAS Viya v1.0.4